+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.2
 
 - Fix default codegen.
+- Call skipWaiting() by default (can be disabled through flag).
 
 ## 0.0.1