+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Fix default codegen.
 - Call skipWaiting() by default (can be disabled through flag).
+- The generated offline URLs and the SW registration is changed to relative URLS,
+  we can put the application in any directory.
 
 ## 0.0.1